+++ tests/testthat/_fixtures/guru/R/paths-apis-get_providers.R
@@ -31,8 +31,5 @@ req_get_providers <- function() {
 }
 
 tidy_policy_get_providers <- function() {
-  spec <- tibblify::tspec_row(
-  tibblify::tib_chr_vec("data", .required = FALSE),
-)
-  nectar::tidy_policy_json_tibblify(spec = spec)
+  nectar::tidy_policy_json(subset_path = "data", simplifyVector = TRUE)
 }