+++ tests/testthat/_fixtures/guru/R/paths-apis-list_apis.R
@@ -31,6 +31,5 @@ req_list_apis <- function() {
 }
 
 tidy_policy_list_apis <- function() {
-  spec <- tibblify::tspec_row()
-  nectar::tidy_policy_json_tibblify(spec = spec)
+  nectar::tidy_policy_json(simplifyVector = TRUE)
 }